+++ app/src/main/java/com/keystone/cold/viewmodel/multisigs/PsbtLegacyConfirmViewModel.java
@@ -548,7 +548,7 @@ public class PsbtLegacyConfirmViewModel extends ParsePsbtViewModel {
                         hdPath = wallet.getExPubPath() + hdPath.substring(1);
                     }
                     utxo.put("publicKey", findMyPubKey(bip32Derivation));
-                    utxo.put("value", psbtInput.optInt("value"));
+                    utxo.put("value", psbtInput.optLong("value"));
                     in.put("utxo", utxo);
                     in.put("ownerKeyPath", hdPath);
                     in.put("masterFingerprint", wallet.getBelongTo());
@@ -598,7 +598,7 @@ public class PsbtLegacyConfirmViewModel extends ParsePsbtViewModel {
                 JSONObject out = new JSONObject();
                 String address = psbtOutput.getString("address");
                 out.put("address", address);
-                out.put("value", psbtOutput.getInt("value"));
+                out.put("value", psbtOutput.getLong("value"));
                 JSONArray bip32Derivation = psbtOutput.optJSONArray("hdPath");
                 if (bip32Derivation != null) {
                     for (int j = 0; j < bip32Derivation.length(); j++) {
